fix(it3): build veri-ezme + tip hataları + UI (sticky sidebar, graph dark colorMode)
- build script: ingest çıkarıldı (her build'de 363->354 ezme bug'ı giderildi)
- tip: vite-env.d.ts (vite/client), vite.config->vitest/config, Icon style/title prop, EChart option cast, schema order int->number (5 float)
- UI: AppShell app-shell düzeni (header sabit, sidebar+main bağımsız scroll -> sidebar STICKY), React Flow colorMode=tema (+/- kontrol/minimap dark)

+++ src/schemas/task.ts
@@ -183,7 +183,7 @@ export const TaskNodeSchema = z
     slug: z.string().min(1),
     summary: z.string().default(""),
     parentId: z.string().nullable().default(null),
-    order: z.number().int().default(0),
+    order: z.number().default(0),
     icon: z.string().default("ph-cube"),
     tags: z.array(z.string()).default([]),
